Added actions to format list
With this, users can define actions to happen when certain result forms
occur.  e.g. you can make it beep when things "fail" or print a message
that everything is okay when things are back to normal.  I may
parameterize this so that you can apply arguments to the action, for
better status messages, e.g. "BOOP-WITH-ID is back to normal"