Added an optional sort function for result format
This function should be a comparator and can be used to reorganise the
dots into some order based on the return status

This is nil by default since I feel part of the usefulness of this comes
from the boops staying in the same order... However, this cannot be
guaranteed when `boop-start` is called anyway due to the asynchronous
nature of the tasks.

It may be good to add in a toggle format to display the order of them
with their ids before returning to the symbols?